Refactor: Moved config parsing to separate function
This is in preparation of making the script an actual module.
While we don't need to parse arguments as part of a module, being able to
parse the config file belonging to the module is handy.
Consequences: Defaults are in get_config(). Code is shorter and cleaner.
In addition, added input validation (Checked that failure_rate is in range
[0,1) ).
Also synchronised the options the command line parser and the config parser
accept. Inlined or localized some single-use functions related to parsing.
Changed the defaults of the command line parser so that they won't interfere
with the config parser if not present.
Finally, changed the type of the collections of URLs from list to set,
since the elements need to be unique and we don't need sequential access

Signed-off-by: Gesh <moystovi@g.jct.ac.il>